fix - apdu: allow the full 254-octet APDU (NPDU length uint16 + FTC send guard 251)
- npdu: length() returns uint16 so octetCount 254 no longer wraps to 0 (frame then dropped by valid())
- bau_systemB: raise ftcSendCommand guard 250 -> 251 (buffer-exact; octetCount 254 is the last valid value, 255 is the escape)

+++ src/knx/npdu.cpp
@@ -18,9 +18,9 @@ void NPDU::octetCount(uint8_t value)
     _data[0] = value;
 }
 
-uint8_t NPDU::length() const
+uint16_t NPDU::length() const
 {
-    return _data[0] + 2; // +1 for length field, +1 for TCPI
+    return (uint16_t)_data[0] + 2; // +1 for length field, +1 for TCPI; uint16 so octetCount 254 -> 256 (no uint8 wrap to 0)
 }
 
 uint8_t NPDU::hopCount() const
